refactor(viewer): split oversized runtime surfaces
