+++ crates/pixel_world_bridge/src/host_state.rs
@@ -2,8 +2,12 @@ use serde_json::{Map, Value, json};
 
 #[path = "host_micro_depot_projection.rs"]
 mod micro_depot_projection;
+#[path = "host_resource_summary.rs"]
+mod resource_summary_projection;
 #[path = "host_social_links.rs"]
 mod social_links;
+
+use resource_summary_projection::{count_resource_entries, resource_summary};
 
 const FRAGMENT_TERRAIN_PALETTE: &[(&str, [u8; 3])] = &[
     ("silicate_matrix", [126, 144, 99]),
@@ -126,40 +130,6 @@ fn world_center_position(world_bounds: &Value) -> Option<Value> {
         "y_cm": number_key(world_bounds, "depth_cm", 0.0) / 2.0,
         "z_cm": number_key(world_bounds, "height_cm", 0.0) / 2.0,
     }))
-}
-
-fn resource_summary(resources: &Value) -> String {
-    let Some(resources) = resources.as_object() else {
-        return "-".to_string();
-    };
-    let entries: Vec<String> = resources
-        .iter()
-        .map(|(key, value)| {
-            if value.is_object() {
-                format!("{key}:{}", value)
-            } else if let Some(text) = value.as_str() {
-                format!("{key}:{text}")
-            } else {
-                format!("{key}:{value}")
-            }
-        })
-        .collect();
-    if entries.is_empty() {
-        "-".to_string()
-    } else {
-        entries.join(" · ")
-    }
-}
-
-fn count_resource_entries(summary: &str) -> usize {
-    if summary.is_empty() || summary == "-" {
-        return 0;
-    }
-    summary
-        .split(" · ")
-        .map(str::trim)
-        .filter(|entry| !entry.is_empty())
-        .count()
 }
 
 fn agent_power_state(agent: &Value) -> Option<&'static str> {
